refactor: clientCourses controllers refactored
