+++ backend/src/controllers/clientCourses/getClientCourse/utils/mappers.ts
@@ -1,0 +1,11 @@
+import { IClientCoursePopulated } from 'interfaces/Ientities/IclientCourses';
+
+const mapClientCourse = (
+  clientCourse: IClientCoursePopulated,
+  similarCourses: never[],
+): IClientCoursePopulated => ({
+  ...clientCourse,
+  course: { ...clientCourse.course, similarCourses },
+});
+
+export { mapClientCourse };
