+++ backend/src/controllers/clientCourses/index.ts
@@ -1,0 +1,6 @@
+export { default as applyCourse } from './applyCourse/applyCourse';
+export { default as getAllClientCourses } from './getAllClientCourses/getAllClientCourses';
+export { default as getClientCourseById } from './getClientCourse/getClientCourse';
+export { default as manageAssessment } from './manageAssessment/manageAssessment';
+export { default as passCourse } from './passCourse/passCourse';
+export { default as startCourse } from './startCourse/startCourse';
